Fix particle id

+++ src/serial/CUDADataFormats/TrackingRecHit2DSOAView.h
@@ -61,8 +61,8 @@ public:
   inline int16_t clusterSizeY(int i) const { return m_ysize[i]; }
   inline uint16_t& detectorIndex(int i) { return m_detInd[i]; }
   inline uint16_t detectorIndex(int i) const { return m_detInd[i]; }
-  inline uint16_t& particleIndex(int i) { return m_pInd[i]; }
-  inline uint16_t particleIndex(int i) const { return m_pInd[i]; }
+  inline int64_t& particleIndex(int i) { return m_pInd[i]; }
+  inline int64_t particleIndex(int i) const { return m_pInd[i]; }
 
   inline pixelCPEforGPU::ParamsOnGPU const& cpeParams() const { return *m_cpeParams; }
 
@@ -97,7 +97,7 @@ private:
   uint16_t* m_detInd;
 
   // supporting objects
-  uint16_t* m_pInd;
+  int64_t* m_pInd;
 
   AverageGeometry*
       m_averageGeometry;  // owned (corrected for beam spot: not sure where to host it otherwise)
